+++ src/js/bun/sql.ts
@@ -1,16 +1,14 @@
 import type * as BunTypes from "bun";
-import type * as QueryTypes from "internal/sql/query.ts";
-import type { SSLMode } from "internal/sql/shared.ts";
-
-type Query<T, Helper extends QueryTypes.BaseQueryHandle> = QueryTypes.Query<T, Helper>;
+import type { PostgresAdapter, PostgresDotZig } from "internal/sql/postgres";
+import type { Query } from "internal/sql/query";
 
 const {
   Query,
   SQLQueryFlags,
   symbols: { _handle, _flags, _results },
 } = require("internal/sql/query");
-const { normalizeQuery } = require("internal/sql/postgres");
-const { SQLHelper, parseOptions, SSLMode, SQLResultArray } = require("internal/sql/shared");
+const { PostgresAdapter } = require("internal/sql/postgres");
+const { SQLHelper, parseOptions, SQLResultArray } = require("internal/sql/shared");
 const { connectionClosedError } = require("internal/sql/utils");
 
 const cmds = ["", "INSERT", "DELETE", "UPDATE", "MERGE", "SELECT", "MOVE", "FETCH", "COPY"];
@@ -23,44 +21,7 @@ const {
   createConnection: createPostgresConnection,
   createQuery: createPostgresQuery,
   init: initPostgres,
-} = $zig("postgres.zig", "createBinding") as {
-  init: (
-    onResolveQuery: (
-      query: Query<any, any>,
-      result: InstanceType<typeof SQLResultArray>,
-      commandTag: string,
-      count: number,
-      queries: any,
-      is_last: boolean,
-    ) => void,
-    onRejectQuery: (query: Query<any, any>, err: Error, queries) => void,
-  ) => void;
-  createConnection: (
-    hostname: string | undefined,
-    port: number,
-    username: string,
-    password: string,
-    databae: string,
-    sslmode: SSLMode,
-    tls: Bun.TLSOptions | boolean | null, // boolean true => empty TLSOptions object `{}`, boolean false or null => nothing
-    query: string,
-    path: string,
-    onConnected: (err: Error | null, connection: $ZigGeneratedClasses.PostgresSQLConnection) => void,
-    onDisconnected: (err: Error | null, connection: $ZigGeneratedClasses.PostgresSQLConnection) => void,
-    idleTimeout: number,
-    connectionTimeout: number,
-    maxLifetime: number,
-    useUnnamedPreparedStatements: boolean,
-  ) => $ZigGeneratedClasses.PostgresSQLConnection;
-  createQuery: (
-    sql: string,
-    values: unknown[],
-    pendingValue: InstanceType<typeof SQLResultArray>,
-    columns: string[] | undefined,
-    bigint: boolean,
-    simple: boolean,
-  ) => $ZigGeneratedClasses.PostgresSQLQuery;
-};
+} = $zig("postgres.zig", "createBinding") as PostgresDotZig;
 
 initPostgres(
   function onResolvePostgresQuery(
@@ -149,615 +110,17 @@ initPostgres(
   },
 );
 
-function onQueryFinish(onClose: (err: Error) => void) {
-  this.queries.delete(onClose);
-  this.pool.release(this);
-}
+// function doCreateQuery(
+//   strings: string | TemplateStringsArray,
+//   values: any[],
+//   allowUnsafeTransaction: boolean,
+//   poolSize: number,
+//   bigint: boolean,
+//   simple: boolean,
+//   adapter: PostgresAdapter,
+// ) {
 
-enum PooledConnectionState {
-  pending = 0,
-  connected = 1,
-  closed = 2,
-}
-
-enum PooledConnectionFlags {
-  /// canBeConnected is used to indicate that at least one time we were able to connect to the database
-  canBeConnected = 1 << 0,
-  /// reserved is used to indicate that the connection is currently reserved
-  reserved = 1 << 1,
-  /// preReserved is used to indicate that the connection will be reserved in the future when queryCount drops to 0
-  preReserved = 1 << 2,
-}
-
-class PooledPostgresConnection {
-  pool: PostgresConnectionPool;
-  connection: $ZigGeneratedClasses.PostgresSQLConnection | null = null;
-  state: PooledConnectionState = PooledConnectionState.pending;
-  storedError: Error | null = null;
-  queries: Set<(err: Error) => void> = new Set();
-  onFinish: ((err: Error | null) => void) | null = null;
-  connectionInfo: any;
-  flags: number = 0;
-  /// queryCount is used to indicate the number of queries using the connection, if a connection is reserved or if its a transaction queryCount will be 1 independently of the number of queries
-  queryCount: number = 0;
-  #onConnected(err, _) {
-    const connectionInfo = this.connectionInfo;
-    if (connectionInfo?.onconnect) {
-      connectionInfo.onconnect(err);
-    }
-    this.storedError = err;
-    if (!err) {
-      this.flags |= PooledConnectionFlags.canBeConnected;
-    }
-    this.state = err ? PooledConnectionState.closed : PooledConnectionState.connected;
-    const onFinish = this.onFinish;
-    if (onFinish) {
-      this.queryCount = 0;
-      this.flags &= ~PooledConnectionFlags.reserved;
-      this.flags &= ~PooledConnectionFlags.preReserved;
-
-      // pool is closed, lets finish the connection
-      // pool is closed, lets finish the connection
-      if (err) {
-        onFinish(err);
-      } else {
-        this.connection?.close();
-      }
-      return;
-    }
-    this.pool.release(this, true);
-  }
-  #onClose(err) {
-    const connectionInfo = this.connectionInfo;
-    if (connectionInfo?.onclose) {
-      connectionInfo.onclose(err);
-    }
-    this.state = PooledConnectionState.closed;
-    this.connection = null;
-    this.storedError = err;
-
-    // remove from ready connections if its there
-    this.pool.readyConnections.delete(this);
-    const queries = new Set(this.queries);
-    this.queries.clear();
-    this.queryCount = 0;
-    this.flags &= ~PooledConnectionFlags.reserved;
-
-    // notify all queries that the connection is closed
-    for (const onClose of queries) {
-      onClose(err);
-    }
-    const onFinish = this.onFinish;
-    if (onFinish) {
-      onFinish(err);
-    }
-
-    this.pool.release(this, true);
-  }
-  constructor(connectionInfo, pool: PostgresConnectionPool) {
-    this.state = PooledConnectionState.pending;
-    this.pool = pool;
-    this.connectionInfo = connectionInfo;
-    this.#startConnection();
-  }
-  async #startConnection() {
-    this.connection = await createConnection(
-      this.connectionInfo,
-      this.#onConnected.bind(this),
-      this.#onClose.bind(this),
-    );
-  }
-  onClose(onClose: (err: Error) => void) {
-    this.queries.add(onClose);
-  }
-
-  bindQuery(query: Query<any, any>, onClose: (err: Error) => void) {
-    this.queries.add(onClose);
-    query.finally(onQueryFinish.bind(this, onClose));
-  }
-
-  #doRetry() {
-    if (this.pool.closed) {
-      return;
-    }
-    // reset error and state
-    this.storedError = null;
-    this.state = PooledConnectionState.pending;
-    // retry connection
-    this.#startConnection();
-  }
-  close() {
-    try {
-      if (this.state === PooledConnectionState.connected) {
-        this.connection?.close();
-      }
-    } catch {}
-  }
-  flush() {
-    this.connection?.flush();
-  }
-  retry() {
-    // if pool is closed, we can't retry
-    if (this.pool.closed) {
-      return false;
-    }
-    // we need to reconnect
-    // lets use a retry strategy
-
-    // we can only retry if one day we are able to connect
-    if (this.flags & PooledConnectionFlags.canBeConnected) {
-      this.#doRetry();
-    } else {
-      // analyse type of error to see if we can retry
-      switch (this.storedError?.code) {
-        case "ERR_POSTGRES_UNSUPPORTED_AUTHENTICATION_METHOD":
-        case "ERR_POSTGRES_UNKNOWN_AUTHENTICATION_METHOD":
-        case "ERR_POSTGRES_TLS_NOT_AVAILABLE":
-        case "ERR_POSTGRES_TLS_UPGRADE_FAILED":
-        case "ERR_POSTGRES_INVALID_SERVER_SIGNATURE":
-        case "ERR_POSTGRES_INVALID_SERVER_KEY":
-        case "ERR_POSTGRES_AUTHENTICATION_FAILED_PBKDF2":
-          // we can't retry these are authentication errors
-          return false;
-        default:
-          // we can retry
-          this.#doRetry();
-      }
-    }
-    return true;
-  }
-}
-
-class PostgresConnectionPool {
-  connectionInfo: Bun.SQL.__internal.DefinedPostgresOptions;
-
-  connections: PooledPostgresConnection[];
-  readyConnections: Set<PooledPostgresConnection>;
-
-  waitingQueue: Array<(err: Error | null, result: any) => void> = [];
-  reservedQueue: Array<(err: Error | null, result: any) => void> = [];
-
-  poolStarted: boolean = false;
-  closed: boolean = false;
-  totalQueries: number = 0;
-  onAllQueriesFinished: (() => void) | null = null;
-
-  constructor(connectionInfo: Bun.SQL.__internal.DefinedPostgresOptions) {
-    this.connectionInfo = connectionInfo;
-    this.connections = new Array(connectionInfo.max);
-    this.readyConnections = new Set();
-  }
-
-  maxDistribution() {
-    if (!this.waitingQueue.length) return 0;
-    const result = Math.ceil((this.waitingQueue.length + this.totalQueries) / this.connections.length);
-    return result ? result : 1;
-  }
-
-  flushConcurrentQueries() {
-    const maxDistribution = this.maxDistribution();
-    if (maxDistribution === 0) {
-      return;
-    }
-
-    while (true) {
-      const nonReservedConnections = Array.from(this.readyConnections).filter(
-        c => !(c.flags & PooledConnectionFlags.preReserved) && c.queryCount < maxDistribution,
-      );
-      if (nonReservedConnections.length === 0) {
-        return;
-      }
-      const orderedConnections = nonReservedConnections.sort((a, b) => a.queryCount - b.queryCount);
-      for (const connection of orderedConnections) {
-        const pending = this.waitingQueue.shift();
-        if (!pending) {
-          return;
-        }
-        connection.queryCount++;
-        this.totalQueries++;
-        pending(null, connection);
-      }
-    }
-  }
-
-  release(connection: PooledPostgresConnection, connectingEvent: boolean = false) {
-    if (!connectingEvent) {
-      connection.queryCount--;
-      this.totalQueries--;
-    }
-    const currentQueryCount = connection.queryCount;
-    if (currentQueryCount == 0) {
-      connection.flags &= ~PooledConnectionFlags.reserved;
-      connection.flags &= ~PooledConnectionFlags.preReserved;
-    }
-    if (this.onAllQueriesFinished) {
-      // we are waiting for all queries to finish, lets check if we can call it
-      if (!this.hasPendingQueries()) {
-        this.onAllQueriesFinished();
-      }
-    }
-
-    if (connection.state !== PooledConnectionState.connected) {
-      // connection is not ready
-      if (connection.storedError) {
-        // this connection got a error but maybe we can wait for another
-
-        if (this.hasConnectionsAvailable()) {
-          return;
-        }
-
-        const waitingQueue = this.waitingQueue;
-        const reservedQueue = this.reservedQueue;
-
-        this.waitingQueue = [];
-        this.reservedQueue = [];
-        // we have no connections available so lets fails
-        for (const pending of waitingQueue) {
-          pending(connection.storedError, connection);
-        }
-        for (const pending of reservedQueue) {
-          pending(connection.storedError, connection);
-        }
-      }
-      return;
-    }
-
-    if (currentQueryCount == 0) {
-      // ok we can actually bind reserved queries to it
-      const pendingReserved = this.reservedQueue.shift();
-      if (pendingReserved) {
-        connection.flags |= PooledConnectionFlags.reserved;
-        connection.queryCount++;
-        this.totalQueries++;
-        // we have a connection waiting for a reserved connection lets prioritize it
-        pendingReserved(connection.storedError, connection);
-        return;
-      }
-    }
-    this.readyConnections.add(connection);
-    this.flushConcurrentQueries();
-  }
-
-  hasConnectionsAvailable() {
-    if (this.readyConnections.size > 0) return true;
-    if (this.poolStarted) {
-      const pollSize = this.connections.length;
-      for (let i = 0; i < pollSize; i++) {
-        const connection = this.connections[i];
-        if (connection.state !== PooledConnectionState.closed) {
-          // some connection is connecting or connected
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  hasPendingQueries() {
-    if (this.waitingQueue.length > 0 || this.reservedQueue.length > 0) return true;
-    if (this.poolStarted) {
-      return this.totalQueries > 0;
-    }
-    return false;
-  }
-  isConnected() {
-    if (this.readyConnections.size > 0) {
-      return true;
-    }
-    if (this.poolStarted) {
-      const pollSize = this.connections.length;
-      for (let i = 0; i < pollSize; i++) {
-        const connection = this.connections[i];
-        if (connection.state === PooledConnectionState.connected) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-  flush() {
-    if (this.closed) {
-      return;
-    }
-    if (this.poolStarted) {
-      const pollSize = this.connections.length;
-      for (let i = 0; i < pollSize; i++) {
-        const connection = this.connections[i];
-        if (connection.state === PooledConnectionState.connected) {
-          connection.connection?.flush();
-        }
-      }
-    }
-  }
-
-  async #close() {
-    let pending;
-    while ((pending = this.waitingQueue.shift())) {
-      pending(connectionClosedError(), null);
-    }
-    while (this.reservedQueue.length > 0) {
-      const pendingReserved = this.reservedQueue.shift();
-      if (pendingReserved) {
-        pendingReserved(connectionClosedError(), null);
-      }
-    }
-
-    const promises: Array<Promise<any>> = [];
-
-    if (this.poolStarted) {
-      this.poolStarted = false;
-      const pollSize = this.connections.length;
-      for (let i = 0; i < pollSize; i++) {
-        const connection = this.connections[i];
-        switch (connection.state) {
-          case PooledConnectionState.pending:
-            {
-              const { promise, resolve } = Promise.withResolvers();
-              connection.onFinish = resolve;
-              promises.push(promise);
-              connection.connection?.close();
-            }
-            break;
-
-          case PooledConnectionState.connected:
-            {
-              const { promise, resolve } = Promise.withResolvers();
-              connection.onFinish = resolve;
-              promises.push(promise);
-              connection.connection?.close();
-            }
-            break;
-        }
-        // clean connection reference
-        // @ts-ignore
-        this.connections[i] = null;
-      }
-    }
-
-    this.readyConnections.clear();
-    this.waitingQueue.length = 0;
-    return Promise.all(promises);
-  }
-
-  async close(options?: { timeout?: number }) {
-    if (this.closed) {
-      return;
-    }
-
-    let timeout = options?.timeout;
-    if (timeout) {
-      timeout = Number(timeout);
-      if (timeout > 2 ** 31 || timeout < 0 || timeout !== timeout) {
-        throw $ERR_INVALID_ARG_VALUE("options.timeout", timeout, "must be a non-negative integer less than 2^31");
-      }
-
-      this.closed = true;
-      if (timeout === 0 || !this.hasPendingQueries()) {
-        // close immediately
-        await this.#close();
-        return;
-      }
-
-      const { promise, resolve } = Promise.withResolvers();
-      const timer = setTimeout(() => {
-        // timeout is reached, lets close and probably fail some queries
-        this.#close().finally(resolve);
-      }, timeout * 1000);
-      timer.unref(); // dont block the event loop
-
-      this.onAllQueriesFinished = () => {
-        clearTimeout(timer);
-        // everything is closed, lets close the pool
-        this.#close().finally(resolve);
-      };
-
-      return promise;
-    } else {
-      this.closed = true;
-      if (!this.hasPendingQueries()) {
-        // close immediately
-        await this.#close();
-        return;
-      }
-
-      // gracefully close the pool
-      const { promise, resolve } = Promise.withResolvers();
-
-      this.onAllQueriesFinished = () => {
-        // everything is closed, lets close the pool
-        this.#close().finally(resolve);
-      };
-
-      return promise;
-    }
-  }
-
-  /**
-   * @param {function} onConnected - The callback function to be called when the connection is established.
-   * @param {boolean} reserved - Whether the connection is reserved, if is reserved the connection will not be released until release is called, if not release will only decrement the queryCount counter
-   */
-  connect(onConnected: (err: Error | null, result: any) => void, reserved: boolean = false) {
-    if (this.closed) {
-      return onConnected(connectionClosedError(), null);
-    }
-
-    if (this.readyConnections.size === 0) {
-      // no connection ready lets make some
-      let retry_in_progress = false;
-      let all_closed = true;
-      let storedError: Error | null = null;
-
-      if (this.poolStarted) {
-        // we already started the pool
-        // lets check if some connection is available to retry
-        const pollSize = this.connections.length;
-        for (let i = 0; i < pollSize; i++) {
-          const connection = this.connections[i];
-          // we need a new connection and we have some connections that can retry
-          if (connection.state === PooledConnectionState.closed) {
-            if (connection.retry()) {
-              // lets wait for connection to be released
-              if (!retry_in_progress) {
-                // avoid adding to the queue twice, we wanna to retry every available pool connection
-                retry_in_progress = true;
-                if (reserved) {
-                  // we are not sure what connection will be available so we dont pre reserve
-                  this.reservedQueue.push(onConnected);
-                } else {
-                  this.waitingQueue.push(onConnected);
-                }
-              }
-            } else {
-              // we have some error, lets grab it and fail if unable to start a connection
-              storedError = connection.storedError;
-            }
-          } else {
-            // we have some pending or open connections
-            all_closed = false;
-          }
-        }
-        if (!all_closed && !retry_in_progress) {
-          // is possible to connect because we have some working connections, or we are just without network for some reason
-          // wait for connection to be released or fail
-          if (reserved) {
-            // we are not sure what connection will be available so we dont pre reserve
-            this.reservedQueue.push(onConnected);
-          } else {
-            this.waitingQueue.push(onConnected);
-          }
-        } else if (!retry_in_progress) {
-          // impossible to connect or retry
-          onConnected(storedError ?? connectionClosedError(), null);
-        }
-        return;
-      }
-      // we never started the pool, lets start it
-      if (reserved) {
-        this.reservedQueue.push(onConnected);
-      } else {
-        this.waitingQueue.push(onConnected);
-      }
-      this.poolStarted = true;
-      const pollSize = this.connections.length;
-      // pool is always at least 1 connection
-      const firstConnection = new PooledPostgresConnection(this.connectionInfo, this);
-      this.connections[0] = firstConnection;
-      if (reserved) {
-        firstConnection.flags |= PooledConnectionFlags.preReserved; // lets pre reserve the first connection
-      }
-      for (let i = 1; i < pollSize; i++) {
-        this.connections[i] = new PooledPostgresConnection(this.connectionInfo, this);
-      }
-      return;
-    }
-    if (reserved) {
-      let connectionWithLeastQueries: PooledPostgresConnection | null = null;
-      let leastQueries = Infinity;
-      for (const connection of this.readyConnections) {
-        if (connection.flags & PooledConnectionFlags.preReserved || connection.flags & PooledConnectionFlags.reserved)
-          continue;
-        const queryCount = connection.queryCount;
-        if (queryCount > 0) {
-          if (queryCount < leastQueries) {
-            leastQueries = queryCount;
-            connectionWithLeastQueries = connection;
-          }
-          continue;
-        }
-        connection.flags |= PooledConnectionFlags.reserved;
-        connection.queryCount++;
-        this.totalQueries++;
-        this.readyConnections.delete(connection);
-        onConnected(null, connection);
-        return;
-      }
-      if (connectionWithLeastQueries) {
-        // lets mark the connection with the least queries as preReserved if any
-        connectionWithLeastQueries.flags |= PooledConnectionFlags.preReserved;
-      }
-      // no connection available to be reserved lets wait for a connection to be released
-      this.reservedQueue.push(onConnected);
-    } else {
-      this.waitingQueue.push(onConnected);
-      this.flushConcurrentQueries();
-    }
-  }
-}
-
-async function createConnection(options, onConnected, onClose) {
-  const {
-    hostname,
-    port,
-    username,
-    tls,
-    query,
-    database,
-    sslMode,
-    idleTimeout = 0,
-    connectionTimeout = 30 * 1000,
-    maxLifetime = 0,
-    prepare = true,
-    path,
-  } = options;
-
-  let password = options.password;
-
-  try {
-    if (typeof password === "function") {
-      password = password();
-      if (password && $isPromise(password)) {
-        password = await password;
-      }
-    }
-
-    return createPostgresConnection(
-      hostname,
-      Number(port),
-      username || "",
-      password || "",
-      database || "",
-      // > The default value for sslmode is prefer. As is shown in the table, this
-      // makes no sense from a security point of view, and it only promises
-      // performance overhead if possible. It is only provided as the default for
-      // backward compatibility, and is not recommended in secure deployments.
-      sslMode || SSLMode.disable,
-      tls || null,
-      query || "",
-      path || "",
-      onConnected,
-      onClose,
-      idleTimeout,
-      connectionTimeout,
-      maxLifetime,
-      !prepare,
-    );
-  } catch (e) {
-    onClose(e);
-    return null;
-  }
-}
-
-function doCreateQuery(
-  strings: string | TemplateStringsArray | import("internal/sql/shared.ts").SQLHelper<any> | Query<any, any>,
-  values: any[],
-  allowUnsafeTransaction: boolean,
-  poolSize: number,
-  bigint: boolean,
-  simple: boolean,
-) {
-  const [sqlString, final_values] = normalizeQuery(strings, values);
-
-  if (!allowUnsafeTransaction) {
-    if (poolSize !== 1) {
-      const upperCaseSqlString = sqlString.toUpperCase().trim();
-      if (upperCaseSqlString.startsWith("BEGIN") || upperCaseSqlString.startsWith("START TRANSACTION")) {
-        throw $ERR_POSTGRES_UNSAFE_TRANSACTION("Only use sql.begin, sql.reserved or max: 1");
-      }
-    }
-  }
-
-  return createPostgresQuery(sqlString, final_values, new SQLResultArray(), undefined, !!bigint, !!simple);
-}
+// }
 
 enum ReservedConnectionState {
   acceptQueries = 1 << 0,
@@ -782,7 +145,12 @@ const SQL: typeof Bun.SQL = function SQL(
   definitelyOptionsButMaybeEmpty: Bun.SQL.Options = {},
 ): Bun.SQL {
   const connectionInfo = parseOptions(stringOrUrlOrOptions, definitelyOptionsButMaybeEmpty);
-  const pool = new PostgresConnectionPool(connectionInfo);
+
+  if (connectionInfo.adapter !== "postgres") {
+    throw new Error("SQLite is unsupported right now");
+  }
+
+  const pool = new PostgresAdapter(connectionInfo);
 
   function onQueryDisconnected(this: Query<any, any>, err: Error) {
     // connection closed mid query this will not be called if the query finishes first
